Small changes for 2.2.6
bot.py - Reenabled web api / web interface
forceafk - Removed commented prints that were used for debugging
Fixed repeat command (when used as a slash command in gen chat or elsewhere, it errors. This small change fixes that)

And other.. im lazy.

+++ cogs/staff/misc/repeat.py
@@ -38,7 +38,11 @@ class repeat(commands.Cog):
             await channel.send(message)
             await ctx.reply("Sent message successfully!", ephemeral=True)
         else:
-            await ctx.message.delete()
+            if ctx.interaction == None:
+                await ctx.message.delete()
+            else:
+                await ctx.reply("You can only use that in staff bot commands!", ephemeral=True)
+
 
 async def setup(bot):
     await bot.add_cog(repeat(bot))
